Struct lrlex::LRNonStreamingLexer

pub struct LRNonStreamingLexer<'lexer, 'input: 'lexer, StorageT> { /* fields omitted */ }

An LRNonStreamingLexer holds a reference to a string and can lex it into lrpar::Lexemes. Although the struct is tied to a single string, no guarantees are made about whether the lexemes are cached or not.