lrlex
Struct LRNonStreamingLexerDef

pub struct LRNonStreamingLexerDef<LexerTypesT: LexerTypes>
where
    usize: AsPrimitive<LexerTypesT::StorageT>,
{ /* private fields */ }

This struct represents, in essence, a .l file in memory. From it one can produce an LRNonStreamingLexer which actually lexes inputs.

Implementations

impl<StorageT: 'static + Debug + Hash + PrimInt + Unsigned, LexerTypesT: LexerTypes<StorageT = StorageT>> LRNonStreamingLexerDef<LexerTypesT>
where usize: AsPrimitive<StorageT>, LexerTypesT::StorageT: TryFrom<usize>,

pub fn new_with_options( s: &str, force_lex_flags: LexFlags, default_lex_flags: LexFlags, ) -> LexBuildResult<LRNonStreamingLexerDef<LexerTypesT>>

pub fn lexer<'lexer, 'input: 'lexer>( &'lexer self, s: &'input str, ) -> LRNonStreamingLexer<'lexer, 'input, LexerTypesT>

Return an LRNonStreamingLexer for the String s that will lex relative to this LRNonStreamingLexerDef.

pub fn lex_flags(&self) -> Option<&LexFlags>

Returns the final LexFlags used for this lex source after all forced and default flags have been resolved.