Struct lrlex::LRNonStreamingLexer

pub struct LRNonStreamingLexer<'lexer, 'input: 'lexer, LexemeT, StorageT: Debug> { /* fields omitted */ }

An LRNonStreamingLexer holds a reference to a string and can lex it into lrpar::Lexemes. Although the struct is tied to a single string, no guarantees are made about whether the lexemes are cached or not.

Implementations

impl<'lexer, 'input: 'lexer, LexemeT: Lexeme<StorageT>, StorageT: Copy + Eq + Debug + Hash + PrimInt + TryFrom<usize> + Unsigned> LRNonStreamingLexer<'lexer, 'input, LexemeT, StorageT>

pub fn new(
s: &'input str,
lexemes: Vec<Result<LexemeT, LexError>>,
newlines: Vec<usize>
) -> LRNonStreamingLexer<'lexer, 'input, LexemeT, StorageT>

Create a new LRNonStreamingLexer that read in: the input s; and derived lexemes and newlines. The newlines Vec<usize> is a sorted list of the byte index of the start of the following line. i.e. for the input string " a\nb\n c d" the Vec should contain [3, 5].

Note that if one or more lexemes or newlines was not created from s, subsequent calls to the LRNonStreamingLexer may cause panics.