lrlex

Struct LRNonStreamingLexer

pub struct LRNonStreamingLexer<'lexer, 'input: 'lexer, LexerTypesT: LexerTypes>
where
    usize: AsPrimitive<LexerTypesT::StorageT>,
    LexerTypesT::StorageT: 'static + Debug + PrimInt,
{ /* private fields */ }

An LRNonStreamingLexer holds a reference to a string and can lex it into lrpar::Lexemes. Although the struct is tied to a single string, no guarantees are made about whether the lexemes are cached or not.

Implementations

impl<'lexer, 'input: 'lexer, StorageT: 'static + Debug + Hash + PrimInt + Unsigned, LexerTypesT: LexerTypes<StorageT = StorageT>> LRNonStreamingLexer<'lexer, 'input, LexerTypesT>
where usize: AsPrimitive<StorageT>,

pub fn new( s: &'input str, lexemes: Vec<Result<LexerTypesT::LexemeT, LRLexError>>, newlines: NewlineCache, ) -> LRNonStreamingLexer<'lexer, 'input, LexerTypesT>

Create a new LRNonStreamingLexer that read in: the input s; and derived lexemes and newlines.

Note that if one or more lexemes or newlines was not created from s, subsequent calls to the LRNonStreamingLexer may cause panics.